Move the Configuration classes to an API module
